+++ packages/manager/modules/vrack/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-vrack@1.5.1...@ovh-ux/manager-vrack@1.5.2) (2023-01-02)
+
+
+### Bug Fixes
+
+* **dedicated.vrack:** load translations ([c5e3324](https://github.com/ovh/manager/commit/c5e33247d2ea6315a43a04bafa0263d8573d7d27))
+
+
+
+
+
 ## [1.5.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-vrack@1.5.0...@ovh-ux/manager-vrack@1.5.1) (2022-11-29)
 
 **Note:** Version bump only for package @ovh-ux/manager-vrack